refactor(atomicwrite): move the staged file write into an internal package
